* 还原使用 zset

+++ runtime/src/main/java/com/easyacc/hutch/publisher/LimitPublisher.java
@@ -37,9 +37,8 @@ public interface LimitPublisher {
     var key = LimitPublisher.zsetKey(hc, msg);
     // TODO: 如果仅仅是使用默认的 Time 作为 score, 那么着就是一个 FIFO/LIFO 的队列, 那么直接使用 LIST 算法上会更快
     Hutch.redis()
-        //    .lpush(key, msg);
         //         使用当前时间作为 score
-        .zadd(key, (double) Timestamp.valueOf(LocalDateTime.now()).getTime(), msg);
+        .zadd(key, Timestamp.valueOf(LocalDateTime.now()).getTime(), msg);
   }
 
   /**
